% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amputeLongData.R
\name{amputeLongData}
\alias{amputeLongData}
\title{amputeLongData}
\usage{
amputeLongData(
  long_data,
  plot.dir = "Missing_Data",
  growth.config = NULL,
  status.config = NULL,
  focus.variable = "SCALE_SCORE",
  student.factors = NULL,
  group = NULL,
  additional.data = NULL,
  compact.results = FALSE,
  ampute.var.weights = NULL,
  reverse.weight = "SCALE_SCORE",
  ampute.args = list(prop = 0.3, type = "RIGHT"),
  complete.cases.only = TRUE,
  partial.fill = TRUE,
  invalidate.repeater.dups = TRUE,
  seed = 4224L,
  R = 10
)
}
\arguments{
\item{long_data}{The complete dataset in which to create missing values.}

\item{plot.dir}{Directory path for diagnostic plots to be placed. Default is
the working directory. Additional subdirectories are created internally as needed.
Default: 'Missing_Data'}

\item{growth.config}{An elongated `SGP` style config script with an entry for
each grade/content_area/year cohort definition. Configs will be used to subset
the `long_data` provided as required for cohort level data amputation.
Default: NULL}

\item{status.config}{An elongated `SGP` style config script with an entry for
each grade/content_area/year cohort definition. Configs will be used to subset
the `long_data` provided as required for cohort level data amputation. Unlike
a `growth.config` entry, `status.config` entries use data from the same grade,
but from the prior year(s) (i.e. not individual variables). For example you
might ampute 3rd grade ELA scores based on a previous year's 3rd grade
school mean scale score, FRL status, etc. Default: NULL}

\item{focus.variable}{The variable to be amputed. Currently only a single
variable is allowed. Default: 'SCALE_SCORE'}

\item{student.factors}{Demographic, (prior) academic achievement or other
student level background information that will be used in the construction
of the weighted scores that define the probability of being missing. Any
i For example, if `c('SCALE_SCORE', 'FREE_REDUCED_LUNCH_STATUS')` is provided,
student level scores and FRL status will be used to calculate weighted scores.
The default (NULL) means that no factors are considered, creating a 'missing
completely at random' (MCAR) missing data pattern. Default: NULL}

\item{group}{Grouping indicator (e.g. institution ID) used to construct
group level means of any `student.factors` provided. Default: NULL}

\item{additional.data}{The function will return only data that is required for
a SGP analysis (based on growth.config). This allows for the addition of more
data (e.g. prior year(s) data not used for data amputation). Default: NULL}

\item{compact.results}{By default (`FALSE`), the function will return a list of
longitudinal datasets with the current (amputed) and prior (unchanged) student
records. This is helpful for diagnostics and ease of use, but also produces more
redundant prior data than needed. Setting this argument to TRUE returns a single
data.table object with a `TRUE/FALSE` indicator column added for each requested
amputation. This flag can be used to make the `SCALE_SCORE` (and/or other variables)
NA in subsequent use cases. Default: `FALSE`}

\item{ampute.var.weights}{Relative weights assigned to the `student.factors`. Default
is NULL meaning the weighted sum scores will be calculated with equal weight (=1).
A named list can be provided with the desired relative weights. For example,
`list(SCALE_SCORE=3, FREE_REDUCED_LUNCH_STATUS=2, SCHOOL_NUMBER=1)` will weight
a student's scale scores by a factor of 3 and FRL by 2, with all (any other `student.factors`)
remaining at the default of 1. This includes school level aggregates in this
example. Note that differential weights for institutions should be placed at
the end of the list. If institution IDs (e.g., `SCHOOL_NUMBER`) are omitted from
the list, the aggregates will be given the same weight as the associated student
level variable. In the given example, `SCALE_SCORE` and school mean scale score
would be given a relative weight of 3. This argument is ignored when
`student.factors = NULL`. Default: NULL}

\item{reverse.weight}{The current default for `ampute.args$type` is `'RIGHT'`, which
means that students with high weighted scores have the highest probability for
amputation. This makes sense for high % FRL schools, but not for high achieving
students and/or students in high achieving schools. This function inverses the
variable(s) individual (and institutional mean) value(s) so that higher weight
is given to lower scores/means. This argument is ignored when `student.factors = NULL`.
Default: 'SCALE_SCORE'}

\item{ampute.args}{Variables to be used in the mice:ampute.continuous function.
Currently only `prop` and `type` can be modified. See ?mice::ampute.continuous
and ?mice::ampute for more information. The `prop` piece is inexact and has required
some modification. Its still imprecise, particularly for values away
from 0.5 (50% missing). Also, the max missingness is 85%, and for that you need
to set prop=0.95 or greater. Note that the prop gives a total proportion missing,
which accounts for missingness already included in the data due to students with
incomplete longitudinal data histories. For example, if a cohort starts with 5%
students missing due to incomplete histories, an additional 25% will be made
missing to get to the 30% (default) missingness. This last point has been dealt
with in some regards with the next argument, which removes these cases first.
Default: list(prop = 0.3, type = 'RIGHT')}

\item{complete.cases.only}{Should cases without the most recent prior and current
score be removed? This removes students with partial longitudinal histories
from the most recent prior (e.g., 2019) to the current year (e.g., 2021),
producing a 'complete' dataset that is easier to interpret. Default: TRUE}

\item{partial.fill}{Should an attempt be made to fill in some of the demographic
and institution ID information based on students previous values? Part of the
process of the amputeScaleScore function is to take the longitudinal data (`long_data`)
and then first widen and then re-lengthen the data, which creates holes for
students with incomplete longitudinal records. This part of the function fills
in these holes for students with existing missing data. Default: TRUE}

\item{invalidate.repeater.dups}{Students who repeat a grade will get missing data
rows inserted for the grade that they 'should' be in for any given year. This leads
to duplicated cases that can lead to problems in the SGP analyses. This argument
returns those cases with the `VALID_CASE` variable set to `'INVALID_CASE'`. Default: TRUE}

\item{seed}{A random seed set for the amputation process to allow for replication
of results, or for alternative results using the same code. Default: 4224}

\item{R}{The number of amputed datasets to return. Default: 10}
}
\value{
Function returns either a list (default) of amputed longitudinal data
 sets or a single data set containing additional columns of indicators for
 records to be removed.
}
\description{
Function removes cases to produce desired patterns and proportions of missing data.
}
\details{
From the amputation process specified, the function returns either a list
 (default) of `R` amputed datasets or a single data set with `R` columns of missing
 record indicators. The datasets will exclude data for students not used in any of
 the specified growth.config or status.config cohorts, unless the additional.data 
 argument has been included.
}
\examples{
 \dontrun{
   data_to_ampute <- SGPdata::sgpData_LONG_COVID

   ###   Read in STEP 0 SGP configuration scripts
   source("SGP_CONFIG/STEP_0/Ampute_2021/Growth.R")
   source("SGP_CONFIG/STEP_0/Ampute_2021/Status.R")

   Test_Data_LONG <-
       amputeScaleScore(
           long_data = data_to_ampute,
           growth.config = growth_config_2021,
           status.config = status_config_2021,
           R = 1)
 }
}
\seealso{
\code{\link[mice]{ampute}}, \code{\link[mice]{ampute.continuous}}
}
\author{
Adam R. VanIwaarden \email{avaniwaarden@nciea.org}
}
\keyword{misc}
\keyword{models}
