% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeData.R
\name{completeData}
\alias{completeData}
\title{FUNCTION_TITLE}
\usage{
completeData(
  long_data,
  shift.group = c("ID", "CONTENT_AREA"),
  shift.period = "YEAR",
  shift.max = 2,
  invariant.vars = NULL,
  conditional.vars = NULL
)
}
\arguments{
\item{long_data}{PARAM_DESCRIPTION}

\item{shift.group}{PARAM_DESCRIPTION, Default: c("ID", "CONTENT_AREA")}

\item{shift.period}{PARAM_DESCRIPTION, Default: 'YEAR'}

\item{shift.max}{PARAM_DESCRIPTION, Default: 2}

\item{invariant.vars}{PARAM_DESCRIPTION, Default: NULL}

\item{conditional.vars}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{J}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{shift}}, \code{\link[data.table]{rbindlist}}
}
\author{
AUTHOR [AUTHOR_2]
}
\keyword{KEYWORD_TERM}
