% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_finalfit.R
\name{extract_fit.fixest}
\alias{extract_fit.fixest}
\title{FUNCTION_TITLE}
\usage{
extract_fit.fixest(
  .data,
  explanatory_name = "explanatory",
  estimate_name = "OR",
  estimate_suffix = "",
  p_name = "p",
  exp = TRUE,
  confint_type = "profile",
  confint_level = 0.95,
  ...
)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{explanatory_name}{PARAM_DESCRIPTION, Default: 'explanatory'}

\item{estimate_name}{PARAM_DESCRIPTION, Default: 'OR'}

\item{estimate_suffix}{PARAM_DESCRIPTION, Default: ''}

\item{p_name}{PARAM_DESCRIPTION, Default: 'p'}

\item{exp}{PARAM_DESCRIPTION, Default: TRUE}

\item{confint_type}{PARAM_DESCRIPTION, Default: 'profile'}

\item{confint_level}{PARAM_DESCRIPTION, Default: 0.95}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}, \code{\link[dplyr]{select}}
}
\author{
Adam R. VanIwaarden \email{avaniwaarden@nciea.org}
}
