% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_finalfit.R
\name{ff_metrics.fixest}
\alias{ff_metrics.fixest}
\title{FUNCTION_TITLE}
\usage{
ff_metrics.fixest(.data)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{AIC}}
 \code{\link[pROC]{are.paired}}, \code{\link[pROC]{auc}}, \code{\link[pROC]{ci}}, \code{\link[pROC]{ci.auc}}, \code{\link[pROC]{ci.coords}}, \code{\link[pROC]{ci.se}}, \code{\link[pROC]{ci.sp}}, \code{\link[pROC]{ci.thresholds}}, \code{\link[pROC]{coords}}, \code{\link[pROC]{cov.roc}}, \code{\link[pROC]{has.partial.auc}}, \code{\link[pROC]{lines.roc}}, \code{\link[pROC]{multiclass.roc}}, \code{\link[pROC]{pROC-package}}, \code{\link[pROC]{plot.ci}}, \code{\link[pROC]{plot.roc}}, \code{\link[pROC]{power.roc.test}}, \code{\link[pROC]{print}}, \code{\link[pROC]{roc}}, \code{\link[pROC]{roc.test}}, \code{\link[pROC]{smooth}}, \code{\link[pROC]{var.roc}}
 \code{\link[finalfit]{metrics_hoslem}}
}
\author{
Adam R. VanIwaarden \email{avaniwaarden@nciea.org}
}
