% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_finalfit.R
\name{glmfixed}
\alias{glmfixed}
\title{FUNCTION_TITLE}
\usage{
glmfixed(.data, dependent, explanatory, fixed_effect = NULL, ...)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{dependent}{PARAM_DESCRIPTION}

\item{explanatory}{PARAM_DESCRIPTION}

\item{fixed_effect}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{formula}}
 \code{\link[fixest]{feglm}}
}
\author{
Adam R. VanIwaarden \email{avaniwaarden@nciea.org}
}
