% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeLongData.R
\name{imputeLongData}
\alias{imputeLongData}
\title{FUNCTION_TITLE}
\usage{
imputeLongData(
  long_data,
  return.all.data = TRUE,
  plot.dir = "Missing_Data",
  status.config = NULL,
  growth.config = NULL,
  focus.variable = "SCALE_SCORE",
  student.factors = NULL,
  group = NULL,
  impute.long = FALSE,
  impute.method = NULL,
  parallel.config = NULL,
  seed = 4224L,
  M = 15,
  maxit = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{long_data}{PARAM_DESCRIPTION}

\item{return.all.data}{PARAM_DESCRIPTION, Default: TRUE}

\item{plot.dir}{PARAM_DESCRIPTION, Default: 'Missing_Data'}

\item{status.config}{PARAM_DESCRIPTION, Default: NULL}

\item{growth.config}{PARAM_DESCRIPTION, Default: NULL}

\item{focus.variable}{PARAM_DESCRIPTION, Default: 'SCALE_SCORE'}

\item{student.factors}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{impute.long}{PARAM_DESCRIPTION, Default: FALSE}

\item{impute.method}{PARAM_DESCRIPTION, Default: NULL}

\item{parallel.config}{PARAM_DESCRIPTION, Default: NULL}

\item{seed}{PARAM_DESCRIPTION, Default: 4224}

\item{M}{PARAM_DESCRIPTION, Default: 15}

\item{maxit}{PARAM_DESCRIPTION, Default: 30}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}, \code{\link[parallel]{mclapply}}, \code{\link[parallel]{makeCluster}}, \code{\link[parallel]{RNGstreams}}, \code{\link[parallel]{clusterApply}}
 \code{\link[data.table]{J}}, \code{\link[data.table]{setattr}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{dcast.data.table}}, \code{\link[data.table]{melt.data.table}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{rbindlist}}
 \code{\link[utils]{head}}
 \code{\link[stats]{formula}}, \code{\link[stats]{setNames}}
 \code{\link[mice]{make.blocks}}, \code{\link[mice]{make.predictorMatrix}}, \code{\link[mice]{mice}}, \code{\link[mice]{ibind}}, \code{\link[mice]{densityplot.mids}}, \code{\link[mice]{complete.mids}}
 \code{\link[callr]{r}}
 \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{dev}}
 \code{\link[graphics]{plot.default}}
}
