% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageLog.R
\name{messageLog}
\alias{messageLog}
\title{A function for creating customized log reports and saving message info.}
\usage{
messageLog(
  log.message,
  log.directory = "Logs",
  logfile = NULL,
  add.date = TRUE,
  appendLF = TRUE
)
}
\arguments{
\item{log.message}{Text string to be exported to a text file log.}

\item{log.directory}{File path where log file should be saved.
Default: 'Logs'}

\item{logfile}{Name of file to be saved.  If NULL, then file will be named
as \code{messageLog_*DATE*.txt}, where \code{*DATE*} is the current date.
Default: NULL}

\item{add.date}{Add the date to a user provided (non-NULL) \code{logfile} name?
Default: TRUE}

\item{appendLF}{Add log message to a file (if it already exists)?
Default: TRUE}
}
\value{
A text file with name provided in "logfile", saved in the "log.directory".
}
\description{
Function produces a text file (custom message log).
}
\details{
Creates (or adds to) a log file with the text given in the "log.message" argument.
}
\author{
Adam R. VanIwaarden \email{avaniwaarden@nciea.org}
}
\keyword{misc}
\keyword{models}
