% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingDataAnalysis.R
\name{missingDataAnalysis}
\alias{missingDataAnalysis}
\title{FUNCTION_TITLE}
\usage{
missingDataAnalysis(
  long_data,
  plot.dir = "Missing_Data",
  status.config = NULL,
  growth.config = NULL,
  focus.variable = "SCALE_SCORE",
  student.factors = NULL,
  factor.labels = NULL,
  group = NULL,
  output.plots = FALSE,
  output.type = "SVG",
  ...
)
}
\arguments{
\item{long_data}{PARAM_DESCRIPTION}

\item{plot.dir}{PARAM_DESCRIPTION, Default: 'Missing_Data'}

\item{status.config}{PARAM_DESCRIPTION, Default: NULL}

\item{growth.config}{PARAM_DESCRIPTION, Default: NULL}

\item{focus.variable}{PARAM_DESCRIPTION, Default: 'SCALE_SCORE'}

\item{student.factors}{PARAM_DESCRIPTION, Default: NULL}

\item{factor.labels}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{output.plots}{PARAM_DESCRIPTION, Default: FALSE}

\item{output.type}{PARAM_DESCRIPTION, Default: 'SVG'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{J}}, \code{\link[data.table]{setattr}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{dcast.data.table}}, \code{\link[data.table]{data.table-package}}
 \code{\link[stats]{setNames}}
 \code{\link[naniar]{mcar_test}}, \code{\link[naniar]{gg_miss_upset}}, \code{\link[naniar]{gg_miss_fct}}
 \code{\link[mice]{md.pattern}}
 \code{\link[finalfit]{missing_plot}}, \code{\link[finalfit]{missing_pairs}}, \code{\link[finalfit]{extract_variable_label}}, \code{\link[finalfit]{summary_factorlist}}, \code{\link[finalfit]{ff_merge}}
 \code{\link[ggplot2]{scale_colour_gradient}}, \code{\link[ggplot2]{scale_manual}}, \code{\link[ggplot2]{labs}}
 \code{\link[tools]{toTitleCase}}
 \code{\link[fixest]{fixef_reexported}}
 \code{\link[utils]{combn}}
 \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{dev2}}, \code{\link[grDevices]{recordPlot}}
 \code{\link[VIM]{mosaicMiss}}, \code{\link[VIM]{spineMiss}}
 \code{\link[svglite]{svglite}}
}
\author{
AUTHOR [AUTHOR_2]
}
\keyword{KEYWORD_TERM}
