% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parMICE.R
\name{parMICE}
\alias{parMICE}
\title{FUNCTION_TITLE}
\usage{
parMICE(
  data,
  m = 5,
  maxit = 5,
  meth,
  bloks,
  frmlas,
  visit = NULL,
  seed = NA,
  nnodes = 5,
  cluster.type = "FORK",
  packages = "mice"
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION, Default: 5}

\item{maxit}{PARAM_DESCRIPTION, Default: 5}

\item{meth}{PARAM_DESCRIPTION}

\item{bloks}{PARAM_DESCRIPTION}

\item{frmlas}{PARAM_DESCRIPTION}

\item{visit}{PARAM_DESCRIPTION, Default: NULL}

\item{seed}{PARAM_DESCRIPTION, Default: NA}

\item{nnodes}{PARAM_DESCRIPTION, Default: 5}

\item{cluster.type}{PARAM_DESCRIPTION, Default: "FORK"}

\item{packages}{PARAM_DESCRIPTION, Default: "mice"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{makeCluster}}, \code{\link[parallel]{RNGstreams}}, \code{\link[parallel]{clusterApply}}
 \code{\link[mice]{mice}}
 \code{\link[abind]{abind}}
}
\author{
AUTHOR [AUTHOR_2]
}
\keyword{KEYWORD_TERM}
